Don't release RDMA zero copy buffers back into pool
They don't have their addresses registered with the internal maps if
they are created by the user and we'd have update the maps as new blocks
come in.
Also, the memory belongs to the user and might be reused, so we can't
keep the pointers and assume they are ours.

osu_latency test now running to completion on a single thread.